+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "Azure Terraform" extension will be documented in thi
 		
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.3.3]
+### Changed
+- Removed Azure Account extension dependency
+  - [#254](https://github.com/Azure/vscode-azureterraform/pull/254)
+- CloudShell functionality is temporarily unavailable
+- Default terminal setting now falls back to integrated terminal when CloudShell is selected
+
 ## [0.3.2]
 ### Fixed
 - fix dependency issues